move changes into a separate file
